avoid some newly introduced g++ warnings

+++ shedskin/lib/builtin/bytes.cpp
@@ -57,7 +57,7 @@ __ss_int bytes::find(__ss_int i, __ss_int a, __ss_int b) {
     __ss_int step = 1;
     slicenr(3, a, b, step, this->__len__());
     for(__ss_int j=a; j<b; j++) {
-        if(unit[j] == i)
+        if(unit[(size_t)j] == i)
             return j;
     }
     return -1;
@@ -85,7 +85,7 @@ __ss_int bytes::rfind(__ss_int i, __ss_int a, __ss_int b) {
     __ss_int step = 1;
     slicenr(3, a, b, step, this->__len__());
     for(__ss_int j=b-1; j>=a; j--) {
-        if(unit[j] == i)
+        if(unit[(size_t)j] == i)
             return j;
     }
     return -1;
